% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ABW.R
\name{ABW}
\alias{ABW}
\title{Calculate the adjusted body weight}
\usage{
ABW(val.IBW = NULL, sex = NULL, len = NULL, Weight = NULL)
}
\arguments{
\item{val.IBW}{ideal body weight}

\item{sex}{1 for male and 0 for female, or "M/m" for male and "F/f" for female}

\item{len}{\itemize{
\item in meters
}}

\item{Weight}{in kg}
}
\description{
Calculate the adjusted body weight
}
