% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_data.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Function for obtaining BOOTSTRAP data}
\usage{
bootstrap_data(fpath)
}
\arguments{
\item{fpath}{path to "bootstrap_results.csv"}
}
\value{
None
}
\description{
Function for obtaining BOOTSTRAP data
}
\examples{
-


}
