% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ABW.R
\name{cov_ABW}
\alias{cov_ABW}
\title{Calculate the adjusted body weight}
\usage{
cov_ABW(val.IBW = NULL, sex = NULL, height = NULL, weight = NULL)
}
\arguments{
\item{val.IBW}{ideal body weight}

\item{sex}{1 for male and 0 for female, or "M/m" for male and "F/f" for female}

\item{height}{specified in meters}

\item{weight}{in kg}
}
\value{
None
}
\description{
Calculate the adjusted body weight
}
\examples{
cov_ABW(val.IBW=66.2, weight=58.6)

}
