% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_BMI.R
\name{cov_BMI}
\alias{cov_BMI}
\title{calculate body mass index}
\usage{
cov_BMI(weight, len)
}
\arguments{
\item{weight}{in kg}

\item{len}{in meters}
}
\value{
None
}
\description{
calculate body mass index
}
\examples{
 None

}
