% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_BSA.R
\name{cov_BSA}
\alias{cov_BSA}
\title{Calculate BSA with different methods}
\usage{
cov_BSA(weight, height, method)
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{method}{method to calculate BSA: mosteller, dubois, haycock}
}
\description{
Calculate BSA with different methods
}
\keyword{BSA,}
\keyword{dubois,}
\keyword{haycock}
\keyword{mosteller,}
