% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_LBW.R
\name{cov_LBW}
\alias{cov_LBW}
\title{Calculate lean body weight}
\usage{
cov_LBW(sex, weight, val.BMI = NULL, len = NULL)
}
\arguments{
\item{sex}{1 for male and 0 for female, or "M/m" for male and "F/f" for female}

\item{weight}{\itemize{
\item specified in kg
}}

\item{val.BMI}{\itemize{
\item specified in kg/m2
}}

\item{len}{\itemize{
\item specified in meters
}}
}
\value{
None
}
\description{
Calculate lean body weight
}
