% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focb_locf.R
\name{focb_locf}
\alias{focb_locf}
\title{Providing a last observation carried forward (LOCF) and first observation carried backward (FOCB) on vector}
\usage{
focb_locf(x, na.value = NA)
}
\arguments{
\item{x}{vector which includes NAs}

\item{na.value}{value that contains the values that are overwritten by LOCF or FOCB}
}
\value{
vector overwritten using LOCF and FOCB
}
\description{
Providing a last observation carried forward (LOCF) and first observation carried backward (FOCB) on vector
}
\examples{
focb_locf(c(NA,NA,1,2,3,5,NA,NA,NA))


}
