% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_data.R
\name{func.vpcdat}
\alias{func.vpcdat}
\title{Load VPC data}
\usage{
func.vpcdat(fpath, dsname = F)
}
\arguments{
\item{fpath}{path to VPC data "folder should include vpc_results.csv"}

\item{dsname}{opzoeken}
}
\value{
vpc datasets
}
\description{
Load VPC data
}
\examples{
-


}
