% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NM_table.R
\name{read_table_nm}
\alias{read_table_nm}
\title{Read NONMEM output tables}
\usage{
read_table_nm(
  file = NULL,
  skip = NULL,
  header = NULL,
  rm_duplicates = FALSE,
  nonmem_tab = TRUE
)
}
\arguments{
\item{file}{path to file}

\item{skip}{\itemize{
\item 
}}

\item{header}{TRUE or FALSE}

\item{rm_duplicates}{TRUE or FALSE}

\item{nonmem_tab}{\itemize{
\item 
}}
}
\value{
none
}
\description{
Read NONMEM output tables
}
\examples{
-

}
