% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_by_id.R
\name{seq_by_id}
\alias{seq_by_id}
\title{Create counter of rows for each ID}
\usage{
seq_by_id(id_col)
}
\arguments{
\item{id_col}{Values from the (patient) ID column}
}
\description{
Create counter of rows for each ID
}
\keyword{seq_by_id}
