% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_by_occ.R
\name{seq_by_occ}
\alias{seq_by_occ}
\title{Create counter of rows for each occasion}
\usage{
seq_by_occ(id_col, occ_col)
}
\arguments{
\item{id_col}{Values from the (patient) ID column}

\item{occ_col}{Values from the occasion (OCC) column}
}
\description{
Create counter of rows for each occasion
}
\keyword{seq_by_occ}
